fix: namespaced faretapdetails
