+++ app/graphql/types/tfl/entities/journey_planner/fare_tap_details_type.rb
@@ -1,0 +1,16 @@
+module Types
+  module Tfl
+    module Entities
+      module JourneyPlanner
+        class FareTapDetailsType < Types::BaseObject
+          field :busRouteId, String, null: true
+          field :hostDeviceType, String, null: true
+          field :modeType, String, null: true
+          field :nationalLocationCode, Integer, null: true
+          field :tapTimestamp, String, null: true
+          field :validationType, String, null: true
+        end
+      end
+    end
+  end
+end
