+++ app/graphql/types/tfl/entities/journey_planner/fare_tap_type.rb
@@ -4,7 +4,7 @@ module Types
       module JourneyPlanner
         class FareTapType < Types::BaseObject
           field :atcoCode, String, null: true
-          field :tapDetails, Types::FareTapDetailsType, null: true
+          field :tapDetails, Types::Tfl::Entities::JourneyPlanner::FareTapDetailsType, null: true
         end
       end
     end
