+++ spec/graphql/types/tfl/entities/journey_planner/fare_tap_details_type_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe Types::FareTapDetailsType do
+describe Types::Tfl::Entities::JourneyPlanner::FareTapDetailsType do
   subject { described_class }
 
   it { is_expected.to have_field(:busRouteId).of_type('String') }
